Move chunk_text_from_file function to text_processing.py
